requirements.txt
appdirs==1.4.3
Django==1.10.6
djangorestframework==3.3.3
djangorestframework-jwt==1.7.2
djangorestframework-bulk==0.2.1
flask==0.11.1
gunicorn==19.6.0
MySQL-python==1.2.5
packaging==16.8
pyparsing==2.2.0
python-socketio==1.4.4
six==1.10.0
socketIO_client==0.7.2
websocket-client==0.37.0
